requirements.txt
flask>=3.0.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
opencv-python>=4.7.0
Pillow>=10.0.0
scikit-learn>=1.3.0
